Updating main to match public (#6)
* Update index.md

Corrected markdown error.

Signed-off-by: Kythera of Anevern <50753925+KytheraOA@users.noreply.github.com>

* Update index.md

Signed-off-by: Kythera of Anevern <50753925+KytheraOA@users.noreply.github.com>

* Numbering correction (#4)

* Update index.md

* Update index.md

* Update index.md

Automatic lists were not working correctly with how I have images inserted for each step. I used the backslash command (example:`1\.`) to escape the automatic lists and rely on manual formatting. 

Probably not the most optimal solution but it works for now.

Source for solution: https://stackoverflow.com/questions/50916253/how-to-prevent-disable-automatic-list-number-incrementing-in-markdown

Signed-off-by: Kythera of Anevern <50753925+KytheraOA@users.noreply.github.com>

---------

Signed-off-by: Kythera of Anevern <50753925+KytheraOA@users.noreply.github.com>

* Create README.md

Signed-off-by: Kythera of Anevern <50753925+KytheraOA@users.noreply.github.com>

* Add readme (#5)

* Update index.md

* Update index.md

* Update index.md

Automatic lists were not working correctly with how I have images inserted for each step. I used the backslash command (example:`1\.`) to escape the automatic lists and rely on manual formatting. 

Probably not the most optimal solution but it works for now.

Source for solution: https://stackoverflow.com/questions/50916253/how-to-prevent-disable-automatic-list-number-incrementing-in-markdown

Signed-off-by: Kythera of Anevern <50753925+KytheraOA@users.noreply.github.com>

---------
Adding a readme to explain what this repo is for.

Signed-off-by: Kythera of Anevern <50753925+KytheraOA@users.noreply.github.com>

---------

Signed-off-by: Kythera of Anevern <50753925+KytheraOA@users.noreply.github.com>